src/datagen/requirements_jaxcfd.txt
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.58.4
h5py==3.14.0
ipykernel==6.29.5
ipython==8.37.0
jax==0.6.2
jax-cfd @ git+https://github.com/google/jax-cfd@0ec4a4c3c000b853485fcdf7f13de05f6f53b788
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
nest-asyncio==1.6.0
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.2.5
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
tree-math==0.2.1
typing_extensions==4.14.0
tzdata==2025.2
wcwidth==0.2.13
xarray==2025.6.1
